Linux (on Windows) is up and running.
Got OSRM to work on the Windows Linux Subsystem. Wrote a couple of shell scripts for Ubuntu that should make installing and running pretty easy. Once I confirm that this works, I'll add to the instructions how to clone the repo I've made and instructions for running the helper scripts.